refactor(connect): 커넥트 쿼리키 분리
Resolve: #330
